Made sure all the working code is in this branch
I made sure the dance code and the code for the brush were all in main. I thought that the dance code in main worked and I also added the working intake command(it had gotten deleted in the recent changes. I also reformated it a little bit by adding intake and outtake commands like we had previoulsy talked about.